# cmd/api/docs/swagger.yaml
definitions:
  domain.Category:
    properties:
      category:
        type: string
      category_image:
        type: string
      id:
        type: integer
    type: object
...
  models.AddToCart:
    properties:
      inventory_id:
        type: integer
      user_id:
        type: integer
    type: object
  models.AdminLogin:
    properties:
      email:
        type: string
      password:
        maxLength: 20
        minLength: 8
        type: string
    required:
      - email
    type: object
...
  models.NewPaymentMethod:
    properties:
      payment_method:
        type: string
    type: object
...
  models.SetNewName:
    properties:
      current:
        type: string
      new:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact: {}
paths:
  /admin/adminlogin:
    post:
      consumes:
        - application/json
      description: Login handler for ahava admins
      parameters:
        - description: Admin login details
          in: body
          name: admin
          required: true
          schema:
            $ref: "#/definitions/models.AdminLogin"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/response.Response"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/response.Response"
      summary: Admin Login
      tags:
        - Admin
  /admin/category:
    delete:
      consumes:
        - application/json
      description: Admin can delete a category
      parameters:
        - description: id
          in: query
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/response.Response"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/response.Response"
      security:
        - Bearer: []
      summary: Delete Category
      tags:
        - Admin
    post:
      consumes:
        - application/json
      description: Admin can add new categories for products
      parameters:
        - description: category
          in: body
          name: category
          required: true
          schema:
            $ref: "#/definitions/domain.Category"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/response.Response"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/response.Response"
      security:
        - Bearer: []
      summary: Add Category
      tags:
        - Admin
    put:
      consumes:
        - application/json
      description: Admin can update name of a category into new name
      parameters:
        - description: set new name
          in: body
          name: set_new_name
          required: true
          schema:
            $ref: "#/definitions/models.SetNewName"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/response.Response"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/response.Response"
      security:
        - Bearer: []
      summary: Update Category
      tags:
        - Admin
...
  /admin/payment/payment-method/new:
    post:
      consumes:
        - application/json
      description: admin can add new payment methods
      parameters:
        - description: payment method
          in: body
          name: payment
          required: true
          schema:
            $ref: "#/definitions/models.NewPaymentMethod"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/response.Response"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/response.Response"
      security:
        - Bearer: []
      summary: ADD NEW PAYMENT METHOD
      tags:
        - Admin
  /admin/users:
    get:
      consumes:
        - application/json
      description: Retrieve users with pagination
      parameters:
        - description: Page number
          in: query
          name: page
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/response.Response"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/response.Response"
      security:
        - Bearer: []
      summary: Get Users
      tags:
        - Admin
  /admin/users/block:
    post:
      consumes:
        - application/json
      description: using this handler admins can block an user
      parameters:
        - description: user-id
          in: query
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/response.Response"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/response.Response"
      security:
        - Bearer: []
      summary: Block User
      tags:
        - Admin
  /admin/users/unblock:
    post:
      consumes:
        - application/json
      description: UnBlock user
      parameters:
        - description: user-id
          in: query
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/response.Response"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/response.Response"
      security:
        - Bearer: []
      summary: UnBlock an existing user
      tags:
        - Admin
...
  /users/check-out:
    get:
      consumes:
        - application/json
      description: Add products to carts  for the purchase
      parameters:
        - description: id
          in: query
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/response.Response"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/response.Response"
      security:
        - Bearer: []
      summary: Checkout section
      tags:
        - User
...
  /users/home/add-to-cart:
    post:
      consumes:
        - application/json
      description: Add products to carts  for the purchase
      parameters:
        - description: Add To Cart
          in: body
          name: cart
          required: true
          schema:
            $ref: "#/definitions/models.AddToCart"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/response.Response"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/response.Response"
      security:
        - Bearer: []
      summary: Add To Cart
      tags:
        - User
...
swagger: "2.0"